requirements.txt
oracledb
streamlit
google-generativeai
pandas
sqlalchemy
fuzzywuzzy
cx_Oracle